server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx[http2]>=0.27.0
cryptography>=42.0.0
